added column provenance to bouteilles

+++ db/migrate/20220330211509_add_prix_to_bouteilles.rb
@@ -1,5 +1,5 @@
 class AddPrixToBouteilles < ActiveRecord::Migration[6.1]
   def change
-    add_column :bouteilles, :prix, :integer
+    add_column :bouteilles, :prix, :float
   end
 end
